tpl/tune handle null value
For providers that use the legacy SDK, after applying the resources, the state contains null values for not specified properties in the config. Those values will be shown as a drift in a `terraform plan`, and can be fixed via `terraform apply -refresh-only`, that they will be assigned with empty value. This mainly because the legacy SDK doesn't differ between null and empty value.

Normally this is not a problem for `tfadd` as it aims to work against the state that is generated via `terraform import`, where there will be no null value. However, as `tfadd` is supposed to extend to support the framework SDK, where it differs the null and empty value. This means in the framework SDK, there might be null value in the state even after `terraform import`. So ultimately we should support it.